revert some test assertions and consolidate error handling into single file
