+++ packages/cli/src/commands/configure/configure_command_factory.test.ts
@@ -20,6 +20,6 @@ void describe('configure command factory', () => {
 
   void it('throws error if no verb subcommand', async () => {
     const output = await commandRunner.runCommand('configure');
-    assert.ok(output.includes('Not enough non-option arguments'));
+    assert.match(output, /Not enough non-option arguments/);
   });
 });
