+++ packages/cli/src/test-utils/command_runner.ts
@@ -1,7 +1,7 @@
 import { Argv } from 'yargs';
 import { AsyncLocalStorage } from 'node:async_hooks';
 import { mock } from 'node:test';
-import { generateCommandFailureHandler } from '../command_failure_handler.js';
+import { generateCommandFailureHandler } from '../error_handler.js';
 
 class OutputInterceptor {
   private output = '';
